Next revision - untested
This revision incorporates many fixes and fixes compared to the alpha
(all the traces were erased and drawn again, because some components had
to be added and some had to be removed). However, this isn't a release -
I haven't ordered these boards yet, I'm now asking people to take a look
at them to minimize the possibility of having wrong boards made. these
will be ordered after the Chinese holidays.

The schematic is now slightly easier to understand, only three
connections are unrouted (and only one actually won't be connected, I'll
look into it but it's fixable by a solder jumper/0R resistor).